Move colorize functions to util folder.
